modifying the integer parameter to long

+++ src/main/java/jpa/MeetingRepository.java
@@ -6,6 +6,6 @@ import jpa.Meeting;
 
 // This will be AUTO IMPLEMENTED by Spring into a Bean
 // CRUD refers Create, Read, Update, Delete
-public interface MeetingRepository extends CrudRepository<Meeting, Integer> {
+public interface MeetingRepository extends CrudRepository<Meeting, Long> {
 
 }
